Fill in proposal : truly cross platform setup
I want to make a blog post about this at some point. I do not have a lot
of time to make it a priority I can give a deadline to. But if we are
missing articles for the advent I'll push this up my list and post it
there instead of my blog.